month-2/capstone/rag-api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "month2-rag-api"
version = "0.1.0"
description = "Production-caliber RAG API with hybrid retrieval, reranking, and evaluation harness."
readme = "../../README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic-settings>=2.2.1",
    "sqlalchemy[asyncio]>=2.0.29",
    "asyncpg>=0.29.0",
    "pgvector>=0.2.5",
    "redis>=5.0.3",
    "httpx>=0.27.0",
    "pwdlib[argon2]>=0.2.0",
    "pyjwt>=2.8.0",
    "python-multipart>=0.0.9",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "opentelemetry-instrumentation-httpx>=0.45b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.45b0",
    "python-json-logger>=2.0.7",
    "alembic>=1.13.1",
    "openai>=1.16.2",
    "anthropic>=0.25.1",
    "ragas>=0.1.7",
    "sentence-transformers>=2.6.1",
    "unstructured[all-docs]>=0.13.0",
    "langchain>=0.1.14",
    "tiktoken>=0.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.6",
    "mypy>=1.9.0",
    "ruff>=0.3.5",
    "httpx>=0.27.0",
    "ipykernel>=6.29.4",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
